reorganize strategies directories
